Move all switch expression recipes to Java 21 migration
