+++ src/main/resources/META-INF/rewrite/java-version-17.yml
@@ -59,9 +59,6 @@ recipeList:
       artifactId: commons-codec
       newVersion: 1.17.x
   - org.openrewrite.java.migrate.AddLombokMapstructBinding
-  - org.openrewrite.java.migrate.lang.SwitchCaseAssignmentsToSwitchExpression
-  - org.openrewrite.java.migrate.lang.SwitchCaseReturnsToSwitchExpression
-  - org.openrewrite.java.migrate.lang.SwitchExpressionYieldToArrow
   - org.openrewrite.java.migrate.UpdateJakartaAnnotationsIfForJavax
   - org.openrewrite.java.migrate.UpdateJakartaAnnotationsIfExistsForJakarta
