+++ src/test/java/org/openrewrite/java/migrate/lang/SwitchCaseReturnsToSwitchExpressionTest.java
@@ -28,7 +28,7 @@ class SwitchCaseReturnsToSwitchExpressionTest implements RewriteTest {
     @Override
     public void defaults(RecipeSpec spec) {
         spec.recipe(new SwitchCaseReturnsToSwitchExpression())
-          .allSources(s -> s.markers(javaVersion(17)));
+          .allSources(s -> s.markers(javaVersion(21)));
     }
 
     @DocumentExample
@@ -304,7 +304,6 @@ class SwitchCaseReturnsToSwitchExpressionTest implements RewriteTest {
     @Test
     void supportMultiLabelWithNullSwitch() {
         rewriteRun(
-          version(
             //language=java
             java(
               """
@@ -327,8 +326,7 @@ class SwitchCaseReturnsToSwitchExpressionTest implements RewriteTest {
                     }
                 }
                 """
-            ),
-            21)
+            )
         );
     }
 }
