+++ src/test/java/org/openrewrite/java/migrate/lang/SwitchExpressionYieldToArrowTest.java
@@ -28,7 +28,7 @@ class SwitchExpressionYieldToArrowTest implements RewriteTest {
     @Override
     public void defaults(RecipeSpec spec) {
         spec.recipe(new SwitchExpressionYieldToArrow())
-          .allSources(s -> s.markers(javaVersion(17)));
+          .allSources(s -> s.markers(javaVersion(21)));
     }
 
     @DocumentExample
@@ -201,7 +201,6 @@ class SwitchExpressionYieldToArrowTest implements RewriteTest {
     @Test
     void supportMultiLabelWithNullSwitch() {
         rewriteRun(
-          version(
             //language=java
             java(
               """
@@ -226,8 +225,7 @@ class SwitchExpressionYieldToArrowTest implements RewriteTest {
                   }
               }
               """
-            ),
-            21)
+            )
         );
     }
 }
